Move whenRevealed ability helper to PlotCard.
